fixes
git-svn-id: https://svn.java.net/svn/electric~svn/trunk@1441 b64ed6c1-75a8-e629-b1da-d0dae74a743a

+++ srcj/com/sun/electric/tool/io/output/PostScript.java
@@ -974,7 +974,7 @@ public class PostScript extends Output
 	{
 		Poly.Type style = poly.getStyle();
 		TextDescriptor td = poly.getTextDescriptor();
-		int size = td.getTrueSize(wnd) * PSSCALE;
+		int size = (int)(td.getTrueSize(wnd) * PSSCALE);
 		Rectangle2D bounds = poly.getBounds2D();
 
 		// get the font size
